check final file inex

+++ lib/GATK/Refine.pm
@@ -134,6 +134,7 @@ sub perform {
     my $recalTable = "${sample_name}${rmdupResultName}${indelResultName}.recal.table";
     my $recalFile = "${sample_name}${rmdupResultName}${indelResultName}.recal.bam";
     my $final_file = "${sample_name}${rmdupResultName}${indelResultName}.recal${baqResultName}.bam";
+    my $final_file_index = change_extension( $final_file, ".bai" );
 
     my $pbs_file = $self->get_pbs_filename( $pbs_dir, $sample_name );
     my $pbs_name = basename($pbs_file);
@@ -145,19 +146,21 @@ sub perform {
 
     my $rmlist       = "";
     my $inputFile    = $sampleFile;
+    my $inputFileIndex    = "${sampleFile}.bai";
     my $resultPrefix = $sample_name;
 
-    my $pbs = $self->open_pbs( $pbs_file, $pbs_desc, $log_desc, $path_file, $result_dir, $final_file, $init_command, 0, $inputFile );
+    my $pbs = $self->open_pbs( $pbs_file, $pbs_desc, $log_desc, $path_file, $result_dir, $final_file_index, $init_command, 0, $inputFile );
 
     if ( !$sorted ) {
       print $pbs "
-if [[ -s $inputFile && ! -s $presortedFile ]]; then
+if [[ (-s $inputFile) && ! (-s $presortedFile) ]]; then
   echo sort=`date` 
   samtools sort -@ $thread -m 4G -o $presortedFile $inputFile
   samtools index $presortedFile
 fi
 ";
       $inputFile = $presortedFile;
+      $inputFileIndex = "${presortedFile}.bai";
       $rmlist    = $rmlist . " $presortedFile ${presortedFile}.bai";
     }
 
@@ -165,12 +168,13 @@ fi
     if ( $remove_duplicate or $mark_duplicate ) {
       my $rmdupFileIndex = change_extension( $rmdupFile, ".bai" );
       print $pbs "
-if [[ -s $inputFile && ! -s $rmdupFile ]]; then
+if [[ (-s $inputFile) && (-s $inputFileIndex) && (! -s $rmdupFile) ]]; then
   echo MarkDuplicates=`date` 
   java $option -jar $picard_jar MarkDuplicates I=$inputFile O=$rmdupFile ASSUME_SORTED=true REMOVE_DUPLICATES=$removeDupLabel CREATE_INDEX=true VALIDATION_STRINGENCY=SILENT M=${rmdupFile}.metrics
 fi
 ";
       $inputFile = $rmdupFile;
+      $inputFileIndex = $rmdupFileIndex;
       $rmlist    = $rmlist . " $rmdupFile $rmdupFileIndex";
     }
 
@@ -178,18 +182,19 @@ fi
       my $indelFileIndex = change_extension( $indelFile, ".bai" );
       my $intervalFile = change_extension( $indelFile, ".intervals" );
       print $pbs "
-if [[ -s $inputFile && ! -s $indelFile ]]; then
+if [[ (-s $inputFile) && (-s $inputFileIndex) && (! -s $indelFile) ]]; then
   echo RealignerTargetCreator=`date` 
   java $option -jar $gatk_jar -T RealignerTargetCreator -nt $thread $fixMisencodedQuals -I $inputFile -R $faFile $indel_vcf -o $intervalFile $restrict_intervals
 fi
 
-if [[ -s $intervalFile && ! -s $indelFile ]]; then
+if [[ (-s $intervalFile) && (! -s $indelFile) ]]; then
   echo IndelRealigner=`date` 
   #InDel parameter referenced: http://www.broadinstitute.org/gatk/guide/tagged?tag=local%20realignment
   java $option -Djava.io.tmpdir=tmpdir -jar $gatk_jar -T IndelRealigner $fixMisencodedQuals -I $inputFile -R $faFile -targetIntervals $intervalFile $indel_vcf --consensusDeterminationModel USE_READS -LOD 0.4 -o $indelFile 
 fi  
 ";
       $inputFile = $indelFile;
+      $inputFileIndex = $indelFileIndex;
       $rmlist    = $rmlist . " $indelFile $indelFileIndex $intervalFile";
     }
 
@@ -197,19 +202,19 @@ fi
     my $printOptions   = "";
 
     print $pbs "
-if [[ -s $inputFile && ! -s $recalTable ]]; then
+if [[ (-s $inputFile) && (-s $inputFileIndex) && (! -s $recalTable) ]]; then
   echo BaseRecalibrator=`date` 
   java $option -jar $gatk_jar -T BaseRecalibrator -nct $thread -rf BadCigar -R $faFile -I $inputFile $knownsitesvcf -o $recalTable $restrict_intervals
 fi
 
-if [[ -s $recalTable && ! -s $recalFile ]]; then
+if [[ (-s $recalTable) && (! -s $recalFile) ]]; then
   echo PrintReads=`date`
   java $option -jar $gatk_jar -T PrintReads --simplifyBAM -nct $thread -rf BadCigar -R $faFile -I $inputFile -BQSR $recalTable -o $recalFile 
 fi
 ";
     if ($baq) {
       print $pbs "
-if [[ -s $recalFile && ! -s $final_file ]]; then
+if [[ (-s $recalFile) && (-s $recalFileIndex) && (! -s $final_file) ]]; then
   echo baq = `date` 
   samtools calmd -Abr $recalFile $faFile > $final_file 
   samtools index $final_file 
@@ -219,7 +224,7 @@ fi
     }
 
     print $pbs "
-if [[ -s $final_file && ! -s ${final_file}.stat ]]; then 
+if [[ (-s $final_file) && (-s $final_file_index) && (! -s ${final_file}.stat) ]]; then 
   echo flagstat = `date` 
   samtools flagstat $final_file > ${final_file}.stat 
   rm $rmlist
